server/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.0
sqlalchemy==2.0.35
pymysql==1.1.1
cryptography==43.0.0
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
bcrypt==4.0.1
pydantic[email]==2.9.0
pydantic-settings==2.5.0
python-multipart==0.0.9
openpyxl==3.1.2
websockets>=12.0
redis>=5.0.0
